refactor: move package manager interaction out of guard
